runtime, runtimepackagestore: Fix 32-bit support detection on client OS with RSAT
The presence of the Get-WindowsFeature cmdlet is not a reliable
indication of a server OS, because that cmdlet is also available on a
client system with the appropriate Remote Server Administration Tools
component installed - but the cmdlet will not work on that system.

Use the cmdlet that is present on all systems instead.

Additionally, warn the user that (at least currently) the .NET Core
64-bit runtime installer is still a 32-bit application, so will not
install on systems with WoW64 disabled
(https://github.com/dotnet/runtime/issues/3087).

Fixes #55.