Make dotnet(core)?-*-windowshosting depend on ANCM range
This essentially turns -windowshosting into a "pinning" package, which
makes sure that ANCM stays within a specific Major.Minor .NET family.
Thanks to this change, users installing dotnetcore-X.Y-windowshosting
on a clean system will get the ANCM from X.Y, not the overall
latest ANCM, making the installation predictable.
The drawback is that only one dotnetcore-X.Y-windowshosting package can
now be installed. Users wanting to upgrade ANCM to a newer .NET release
will need to uninstall the existing -windowshosting first (which will
not change the configuration of the system, since -windowshosting is a
metapackage).